Added a readme
All the source files and Makefile for your project. I expect these to be
based upon your submission for project 2 or 3.

A readme.txt file that answers the following questions, with at least
one complete sentence each:
- Which feature(s) did you implement?
- What steps should a user, such as the person grading your submission,
take in order to see your feature? What specifically should they look
for in the output?